+++ weldx/asdf/util.py
@@ -1,3 +1,4 @@
+"""Utilities for asdf files."""
 from io import BytesIO
 from pathlib import Path
 from typing import Tuple
@@ -8,11 +9,20 @@ from boltons.iterutils import get_path
 
 from weldx.asdf.extension import WeldxAsdfExtension, WeldxExtension
 
-# TODO: these functions be generalized and be public
+__all__ = [
+    "read_buffer",
+    "write_buffer",
+    "write_read_buffer",
+    "asdf_json_repr",
+    "notebook_fileprinter",
+]
+
 # asdf read/write debug tools functions ---------------------------------------
 
 
-def _write_buffer(tree: dict, asdffile_kwargs: dict = None, write_kwargs: dict = None):
+def _write_buffer(
+    tree: dict, asdffile_kwargs: dict = None, write_kwargs: dict = None
+) -> BytesIO:
     """Write ASDF file into buffer.
 
     Parameters
@@ -20,15 +30,16 @@ def _write_buffer(tree: dict, asdffile_kwargs: dict = None, write_kwargs: dict =
     tree:
         Tree object to serialize.
     asdffile_kwargs
-        Additional keywords to pass to asdf.AsdfFile()
+        Additional keywords to pass to `asdf.AsdfFile`
     write_kwargs
-        Additional keywords to pass to asdf.AsdfFile.write_to()
+        Additional keywords to pass to `asdf.AsdfFile.write_to`
         Weldx-Extensions are always set.
 
     Returns
     -------
-    BytesIO
+    io.BytesIO
         Bytes buffer of the ASDF file.
+
     """
     if asdffile_kwargs is None:
         asdffile_kwargs = {}
@@ -49,11 +60,11 @@ def _read_buffer(buffer: BytesIO, open_kwargs: dict = None):
 
     Parameters
     ----------
-    buffer
+    buffer : io.BytesIO
         Buffer containing ASDF file contents
     open_kwargs
-        Additional keywords to pass to asdf.AsdfFile.open()
-        Extensions are always set, copy_arrays=True is set by default.
+        Additional keywords to pass to `asdf.AsdfFile.open`
+        Extensions are always set, ``copy_arrays=True`` is set by default.
 
     Returns
     -------
@@ -78,25 +89,33 @@ def _write_read_buffer(
     tree: dict, asdffile_kwargs=None, write_kwargs=None, open_kwargs=None
 ):
     """Perform a buffered write/read roundtrip of a tree using default ASDF settings.
+
     Parameters
     ----------
     tree
         Tree object to serialize.
     asdffile_kwargs
-        Additional keywords to pass to asdf.AsdfFile()
+        Additional keywords to pass to `asdf.AsdfFile`
     write_kwargs
-        Additional keywords to pass to asdf.AsdfFile.write_to()
+        Additional keywords to pass to `asdf.AsdfFile.write_to`
         Extensions are always set.
     open_kwargs
-        Additional keywords to pass to asdf.AsdfFile.open()
-        Extensions are always set, copy_arrays=True is set by default.
+        Additional keywords to pass to `asdf.AsdfFile.open`
+        Extensions are always set, ``copy_arrays=True`` is set by default.
+
     Returns
     -------
     dict
+
     """
     buffer = _write_buffer(tree, asdffile_kwargs, write_kwargs)
     return _read_buffer(buffer, open_kwargs)
 
+
+# make read/write buffer functions public
+write_buffer = _write_buffer
+read_buffer = _read_buffer
+write_read_buffer = _write_read_buffer
 
 try:  # pragma: no cover
     import IPython
@@ -138,7 +157,7 @@ else:  # pragma: no cover
         Parameters
         ----------
         file
-            filename or BytesIO buffer of ASDF file
+            filename or ``BytesIO`` buffer of ASDF file
         lexer
             Syntax style to use